[FEATURE] [MER-3415] Changing number of attempts available in simple author/flowchart tool (#5157)
The simple author currently defaults to 3 attempts for a screen, with no option to change this. The requirement was to introduce the ability to change the number of allowed attempts. The default should remain at three, but users should be able to specify a different number of attempts (1, 2, 3, 4, or 5). The system should honor the specified number.