[BUG FIX] [MER-4132] check authorProject status == accepted on collaborators (#5393)
When an author accepts a collaborator invitation, the accepted project is not listed in the "Course Author" workspace. So, the collaborator could access the project only by the email invitation link or by hardcoding the URL.

Root cause
The query fetching all the author's projects (as an owner or as a collaborator) incorrectly checked the AuthorProject status and author_id on the owner binding instead of doing it on the collaborator binding.

Fixed issue example (inviting a_new_author@test.com as a collaborator to the Custom Chemistry project)
Note that the project is not immediately listed after the invitation was sent. Only after the collaborator accepts the invitation the corresponding project gets listed.