[enhance] Adds file upload to LogicLab authoring.
In order to better handle LogicLab activity content, this adds a file input element so that content authors can upload new activities generated in other tools instead of submitting them to the developers to manually add to the database.
It adds a stub for getting the maximum points available based on the selected activity.
It adds a toggle so authors can override a derivation activities tutoring flag which removes the need to have two copies of select activities in the database.