[ENHANCEMENT] [MER-4957] Enhance LogicLab authoring and register outOf points
In order to better handle LogicLab activity content, this adds a file input element so that content authors can upload new activities generated in other tools instead of submitting them to the developers to manually add to the database.
It adds a stub for getting the maximum points available based on the selected activity.
It adds a toggle so authors can override a derivation activities tutoring flag which removes the need to have two copies of select activities in the database.

* [enhance] sets model.authoring.parts[o].outOf on activity update.

* Auto format

* use torus utilities for safely managing MathJax typesetting promise

* define utility function for safe async typesetting

* limit math typesetting to preview div

* [fix] addresses a bunch of ai review issues.

In useLabServer, wrap fetch in useEffect.
Use union of strings for ActivitySpecification instead of string.
Cleans up a bunch of fetch calls and potential data leaks.
Uses provided typeset function.
Wraps some form elements in fieldsets.
Adds placeholder element when loading activities.
Addresses a bunch of accessability issues.

* [fix] Adds async error reporting.

Removes lab server probe code.
Uses useLabServer instead of getLabServer.

---------

Co-authored-by: Anders Weinstein <andersw@andrew.cmu.edu>