[BUG FIX] [MER-5012] prevent user from clicking blank placeholder choice in dropdowns (#6020)
https://eliterate.atlassian.net/browse/MER-5012
This disables the initial blank choice that is always added as initial default to a dropdown menu in a multi-input question so that it cannot be selected by the mouse.

On a practice page, when "submitPerPart" option is set, this suffices to prevent submitting the blank choice entirely. This is because in this case there is no "Submit" button and student must select another choice to get question to auto-submit. This is perhaps the most common option on practice pages.

Note, however, the initial default selection of the blank choice can still be submitted in the following cases:

on a practice page, if "submitPerPart" is NOT set, and student clicks the Submit button before selecting any other choice; and
on a scored page, if one submits the page before selecting any other choice, effectively leaving the question blank.
It seems unlikely we want to block 2; torus has always left students free to leave a question blank. Arguably the problem here is that torus doesn't allow authors to provide custom feedback for the special condition of leaving a question blank. But custom feedback less important on scored pages.

It was a little unclear if the ticket required blocking 1. It seems to be implied, but the ticket was focused on clicking the blank choice and there was no response to the comment pointing out this case and no design suggestion for what the desired behavior should be, and we do allow leaving a numeric input blank and submitting. So this PR did not attempt to block 1.

---------

Co-authored-by: Anders Weinstein <andersw@andrew.cmu.edu>